Move ThreadInfo, MacSoundCard, and MacOSThread to oshi-common; set up macOS packages (#3144)
