+++ oshi-common/src/main/java/oshi/driver/common/mac/package-info.java
@@ -1,0 +1,8 @@
+/*
+ * Copyright 2026 The OSHI Project Contributors
+ * SPDX-License-Identifier: MIT
+ */
+/**
+ * Provides functions to query Mac info
+ */
+package oshi.driver.common.mac;
