+++ oshi-common/src/main/java/oshi/software/common/os/mac/package-info.java
@@ -1,0 +1,8 @@
+/*
+ * Copyright 2026 The OSHI Project Contributors
+ * SPDX-License-Identifier: MIT
+ */
+/**
+ * Provides information about Software and OS on macOS
+ */
+package oshi.software.common.os.mac;
