+++ oshi-core-java25/src/main/java/oshi/software/os/mac/MacOSProcessFFM.java
@@ -10,9 +10,9 @@ import static java.lang.foreign.ValueLayout.JAVA_BYTE;
 import static java.lang.foreign.ValueLayout.JAVA_INT;
 import static java.lang.foreign.ValueLayout.JAVA_LONG;
 import static oshi.ffm.mac.MacSystem.GROUP;
-import static oshi.ffm.mac.MacSystem.PASSWD;
 import static oshi.ffm.mac.MacSystem.MAXCOMLEN;
 import static oshi.ffm.mac.MacSystem.MAXPATHLEN;
+import static oshi.ffm.mac.MacSystem.PASSWD;
 import static oshi.ffm.mac.MacSystem.PBI_COMM;
 import static oshi.ffm.mac.MacSystem.PBI_FLAGS;
 import static oshi.ffm.mac.MacSystem.PBI_GID;
@@ -83,11 +83,12 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import oshi.annotation.concurrent.ThreadSafe;
-import oshi.driver.mac.ThreadInfo;
+import oshi.driver.common.mac.ThreadInfo;
 import oshi.ffm.ForeignFunctions;
 import oshi.ffm.mac.IOKit.IOIterator;
 import oshi.ffm.mac.IOKit.IORegistryEntry;
 import oshi.software.common.AbstractOSProcess;
+import oshi.software.common.os.mac.MacOSThread;
 import oshi.software.os.OSThread;
 import oshi.util.GlobalConfig;
 import oshi.util.ParseUtil;
